pyproject.toml
[project]
name = "coin-report"
version = "0.1.0"
description = ""
authors = [
    {name = "Mateus Dias",email = "mateusdias1409@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "resend (>=2.19.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "coin_report", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = "python src/coin_report/app.py"